lm4f: Update examples to the new GPIO API
We updated the GPIO API to use the AHB bus; however the AHP aperture for
GPIO ports A through J needs to be explicitly enabled at runtime. Accessing
the AHB aperture otherwise hardfaults.

To make the examples work again, we call gpio_enable_ahb_aperture() at the
start of main().

Since we're at it, we also take out the ugly register accesses in favor
of the new gpio functions.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>