[stm32f429i-discovery] Added usbmidi example.
This board connects the USB HS interface to the micro usb connector as
opposed to the USB FS interface on the stm32f4-discovery board. This is
why we are using different pins, different periph and different driver.
At the end it should(TM) behave the same as the HS interface implements
FS too.