tests: gadget0: add stm32l1 target
As with F103 generic, there's no readily available L1 board with USB device
A custom LED is used to track boot process, but otherwise this should be
portable to any L1 board, except for the openocd configuration file.

Tests pass straight away, which is good, as it uses the existing f103 usb
driver.